+++ src/app/shared/models/org-card.model.ts
@@ -1,0 +1,11 @@
+export interface orgCard {
+  title: string;
+  owner: string;
+  city: string;
+  ownership: string;
+  price: number | boolean;
+  rate: string; //temp
+  ageFrom: number;
+  ageTo: number;
+  category: string[];
+}
