+++ src/app/shared/organization-card/organization-card.component.scss
@@ -67,5 +67,6 @@
     align-items: flex-start;
     justify-content: flex-start;
     padding-bottom: 15px;
+    min-height: 120px;
   }
 }
