+++ src/app/shared/store/filter.state.ts
@@ -1,9 +1,9 @@
 import { Injectable } from '@angular/core';
-import { State, Action, StateContext } from '@ngxs/store';
+import { State, Action, StateContext, Selector } from '@ngxs/store';
 import { tap } from 'rxjs/operators';
 
-import { orgCard } from 'src/app/mock-org-cards';
-import { OrgCardsService } from 'src/app/shared/services/org-cards.service';
+import { OrgCardsService } from 'src/app/shared/services/org-cards/org-cards.service';
+import { orgCard } from '../models/org-card.model';
 import { setMinAge, setMaxAge, getCards } from './filter.actions';
 
 export interface FilterStateModel {
@@ -35,16 +35,23 @@ export interface FilterStateModel {
 @Injectable()
 export class FilterState {
 
+  @Selector()
+  static orgCards(state: FilterStateModel) {
+    return state.organizationCards
+  }
+
   constructor(private cardsService: OrgCardsService){}
 
   @Action(setMinAge)
     setMinAge({ patchState }: StateContext<FilterStateModel>, { payload }: setMinAge): void {
       patchState({ ageFrom: payload })
     }
+
   @Action(setMaxAge)
     setMaxAge({ patchState }: StateContext<FilterStateModel>, { payload }: setMaxAge): void {
       patchState({ ageTo: payload })
     }
+
   @Action(getCards)
     getCards({ patchState }: StateContext<FilterStateModel>) {
       return this.cardsService.getCards().pipe(
